+++ modules/recheck-core/shared/src/test/scala/codes/quine/labo/recheck/diagnostics/AttackPatternSuite.scala
@@ -11,10 +11,16 @@ class AttackPatternSuite extends munit.FunSuite {
 
   test("AttackPattern#toString") {
     val attack1 = AttackPattern(Seq((UString("ab"), UString("cd"), 2)), UString("ef"), 1)
-    assertEquals(attack1.toString, "'ab' 'cd'³ 'ef'")
+    assertEquals(attack1.toString(AttackPattern.JavaScript), "'ab' + 'cd'.repeat(3) + 'ef'")
+    assertEquals(attack1.toString(AttackPattern.Superscript), "'ab' 'cd'³ 'ef'")
     val attack2 = AttackPattern(Seq((UString(""), UString("cd"), 2)), UString("ef"), 1)
-    assertEquals(attack2.toString, "'cd'³ 'ef'")
+    assertEquals(attack2.toString(AttackPattern.JavaScript), "'cd'.repeat(3) + 'ef'")
+    assertEquals(attack2.toString(AttackPattern.Superscript), "'cd'³ 'ef'")
     val attack3 = AttackPattern(Seq((UString("ab"), UString("cd"), 2)), UString(""), 1)
-    assertEquals(attack3.toString, "'ab' 'cd'³")
+    assertEquals(attack3.toString(AttackPattern.JavaScript), "'ab' + 'cd'.repeat(3)")
+    assertEquals(attack3.toString(AttackPattern.Superscript), "'ab' 'cd'³")
+
+    // The default style is `JavaScript`.
+    assertEquals(attack1.toString, "'ab' + 'cd'.repeat(3) + 'ef'")
   }
 }